requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
